+++ core/config_manager.py
@@ -52,7 +52,7 @@ class ConfigManager:
             "cookies_from_browser": "None",
             "gallery_cookies_from_browser": "None",
             "js_runtime_path": "",  # New setting for JavaScript runtime
-            "output_template": "%(title)s [%(uploader)s][%(upload_date>%m-%d-%Y)s][%(id)s].%(ext)s",
+            "output_template": "%(title)s [%(uploader)s][%(release_date>%m-%d-%Y)s][%(id)s].%(ext)s",
             "embed_chapters": "True",
             "subtitles_embed": "False",
             "subtitles_write": "False",
@@ -87,7 +87,7 @@ class ConfigManager:
             "cookies_from_browser": "None",
             "gallery_cookies_from_browser": "None",
             "js_runtime_path": "",  # New setting for JavaScript runtime
-            "output_template": "%(title)s [%(uploader)s][%(upload_date>%m-%d-%Y)s][%(id)s].%(ext)s",
+            "output_template": "%(title)s [%(uploader)s][%(release_date>%m-%d-%Y)s][%(id)s].%(ext)s",
             "embed_chapters": "True",
             "subtitles_embed": "False",
             "subtitles_write": "False",